+++ src/num/float/wrapper/shared_series.rs
@@ -7,8 +7,6 @@
 // - CONST tables
 // - fn helpers
 
-#[allow(unused_imports)]
-use super::super::shared_docs::*;
 use crate::{Float, cfor, is, paste};
 
 /// Implements methods independently of any features
@@ -40,7 +38,7 @@ macro_rules! impl_float_shared_series {
             /// `exp` and `ln` functions.
             ///
             /// # Formulation
-            #[doc = FORMULA_POWF_SERIES!()]
+            #[doc = crate::_FLOAT_FORMULA_POWF_SERIES!()]
             ///
             /// See also [`ln_series_terms`][Self::ln_series_terms].
             ///
@@ -62,7 +60,7 @@ macro_rules! impl_float_shared_series {
             /// Computes the exponential function $e^x$ using Taylor series expansion.
             ///
             /// # Formulation
-            #[doc = FORMULA_EXP_SERIES!()]
+            #[doc = crate::_FLOAT_FORMULA_EXP_SERIES!()]
             ///
             /// See also [`exp_series_terms`][Self::exp_series_terms].
             pub const fn exp_series(self, terms: $ue) -> Float<$f> {
@@ -87,7 +85,7 @@ macro_rules! impl_float_shared_series {
             /// Calculates $ e^x - 1 $ using the Taylor series expansion.
             ///
             /// # Formulation
-            #[doc = FORMULA_EXP_M1_SERIES!()]
+            #[doc = crate::_FLOAT_FORMULA_EXP_M1_SERIES!()]
             ///
             /// See also [`exp_series_terms`][Self::exp_series_terms].
             pub const fn exp_m1_series(self, terms: $ue) -> Float<$f> {
@@ -111,7 +109,7 @@ macro_rules! impl_float_shared_series {
             /// Calculates $ 2^x $ using the Taylor series expansion.
             ///
             /// # Formulation
-            #[doc = FORMULA_EXP2_SERIES!()]
+            #[doc = crate::_FLOAT_FORMULA_EXP2_SERIES!()]
             ///
             /// The maximum values with a representable result are:
             /// 127 for `f32` and 1023 for `f64`.
@@ -139,7 +137,7 @@ macro_rules! impl_float_shared_series {
             /// small or too big could be impractical to calculate with precision.
             ///
             /// # Formulation
-            #[doc = FORMULA_LN_SERIES!()]
+            #[doc = crate::_FLOAT_FORMULA_LN_SERIES!()]
             ///
             /// See also [`ln_series_terms`][Self::ln_series_terms].
             pub const fn ln_series(self, terms: $ue) -> Float<$f> {
@@ -190,7 +188,7 @@ macro_rules! impl_float_shared_series {
             /// Computes the logarithm to the given `base` using the change of base formula.
             ///
             /// # Formulation
-            #[doc = FORMULA_LOG_SERIES!()]
+            #[doc = crate::_FLOAT_FORMULA_LOG_SERIES!()]
             ///
             /// See also [`ln_series_terms`][Self::ln_series_terms].
             pub const fn log_series(self, base: $f, terms: $ue) -> Float<$f> {
@@ -209,7 +207,7 @@ macro_rules! impl_float_shared_series {
             /// Computes the base-2 logarithm using the change of base formula.
             ///
             /// # Formulation
-            #[doc = FORMULA_LOG2_SERIES!()]
+            #[doc = crate::_FLOAT_FORMULA_LOG2_SERIES!()]
             ///
             /// See also [`ln_series_terms`][Self::ln_series_terms].
             pub const fn log2_series(self, terms: $ue) -> Float<$f> {
@@ -219,7 +217,7 @@ macro_rules! impl_float_shared_series {
             /// Computes the base-10 logarithm using the change of base formula.
             ///
             /// # Formulation
-            #[doc = FORMULA_LOG10_SERIES!()]
+            #[doc = crate::_FLOAT_FORMULA_LOG10_SERIES!()]
             ///
             /// See also [`ln_series_terms`][Self::ln_series_terms].
             pub const fn log10_series(self, terms: $ue) -> Float<$f> {
@@ -237,7 +235,7 @@ macro_rules! impl_float_shared_series {
             /// The sine calculated using Taylor series expansion.
             ///
             /// # Formulation
-            #[doc = FORMULA_SIN_SERIES!()]
+            #[doc = crate::_FLOAT_FORMULA_SIN_SERIES!()]
             ///
             /// This Taylor series converges relatively quickly and uniformly
             /// over the entire domain.
@@ -264,7 +262,7 @@ macro_rules! impl_float_shared_series {
             /// Computes the cosine using taylor series expansion.
             ///
             /// # Formulation
-            #[doc = FORMULA_COS_SERIES!()]
+            #[doc = crate::_FLOAT_FORMULA_COS_SERIES!()]
             ///
             /// This Taylor series converges relatively quickly and uniformly
             /// over the entire domain.
@@ -293,7 +291,7 @@ macro_rules! impl_float_shared_series {
             /// Computes the tangent using Taylor series expansion of sine and cosine.
             ///
             /// # Formulation
-            #[doc = FORMULA_TAN_SERIES!()]
+            #[doc = crate::_FLOAT_FORMULA_TAN_SERIES!()]
             ///
             /// The tangent function has singularities and is not defined for
             /// `cos(x) = 0`. This function clamps `self` within an appropriate range
@@ -319,7 +317,7 @@ macro_rules! impl_float_shared_series {
             /// Computes the arcsine using Taylor series expansion.
             ///
             /// # Formulation
-            #[doc = FORMULA_ASIN_SERIES!()]
+            #[doc = crate::_FLOAT_FORMULA_ASIN_SERIES!()]
             ///
             /// asin is undefined for $ |x| > 1 $ and in that case returns `NaN`.
             ///
@@ -373,7 +371,7 @@ macro_rules! impl_float_shared_series {
             /// Computes the arccosine using the Taylor expansion of arcsine.
             ///
             /// # Formulation
-            #[doc = FORMULA_ACOS_SERIES!()]
+            #[doc = crate::_FLOAT_FORMULA_ACOS_SERIES!()]
             ///
             /// See the [`asin_series_terms`][Self#method.asin_series_terms] table for
             /// information about the number of `terms` needed.
@@ -394,7 +392,7 @@ macro_rules! impl_float_shared_series {
             /// Computes the arctangent using Taylor series expansion.
             ///
             /// # Formulation
-            #[doc = FORMULA_ATAN_SERIES!()]
+            #[doc = crate::_FLOAT_FORMULA_ATAN_SERIES!()]
             ///
             /// The series converges more slowly near the edges of the domain
             /// (i.e., as `self` approaches -1 or 1). For more accurate results,
@@ -457,7 +455,7 @@ macro_rules! impl_float_shared_series {
             /// via the exponent formula.
             ///
             /// # Formulation
-            #[doc = FORMULA_SINH_SERIES!()]
+            #[doc = crate::_FLOAT_FORMULA_SINH_SERIES!()]
             ///
             /// See the [`exp_series_terms`][Self#method.exp_series_terms] table for
             /// information about the number of `terms` needed.
@@ -469,7 +467,7 @@ macro_rules! impl_float_shared_series {
             /// via the exponent formula.
             ///
             /// # Formulation
-            #[doc = FORMULA_COSH_SERIES!()]
+            #[doc = crate::_FLOAT_FORMULA_COSH_SERIES!()]
             ///
             /// See the [`exp_series_terms`][Self#method.exp_series_terms] table for
             /// information about the number of `terms` needed.
@@ -481,7 +479,7 @@ macro_rules! impl_float_shared_series {
             /// hyperbolic sine and cosine.
             ///
             /// # Formulation
-            #[doc = FORMULA_TANH_SERIES!()]
+            #[doc = crate::_FLOAT_FORMULA_TANH_SERIES!()]
             ///
             /// See the [`exp_series_terms`][Self#method.exp_series_terms] table for
             /// information about the number of `terms` needed.
@@ -494,7 +492,7 @@ macro_rules! impl_float_shared_series {
             /// Computes the inverse hyperbolic sine using the natural logarithm definition.
             ///
             /// # Formulation
-            #[doc = FORMULA_ASINH_SERIES!()]
+            #[doc = crate::_FLOAT_FORMULA_ASINH_SERIES!()]
             ///
             /// See also [`ln_series_terms`][Self::ln_series_terms].
             pub const fn asinh_series(self, terms: $ue) -> Float<$f> {
@@ -505,7 +503,7 @@ macro_rules! impl_float_shared_series {
             /// Computes the inverse hyperbolic cosine using the natural logarithm definition.
             ///
             /// # Formulation
-            #[doc = FORMULA_ACOSH_SERIES!()]
+            #[doc = crate::_FLOAT_FORMULA_ACOSH_SERIES!()]
             ///
             /// See also [`ln_series_terms`][Self::ln_series_terms].
             pub const fn acosh_series(self, terms: $ue) -> Float<$f> {
@@ -520,7 +518,7 @@ macro_rules! impl_float_shared_series {
             /// Computes the inverse hyperbolic tangent using the natural logarithm definition.
             ///
             /// # Formulation
-            #[doc = FORMULA_ATANH_SERIES!()]
+            #[doc = crate::_FLOAT_FORMULA_ATANH_SERIES!()]
             ///
             /// See also [`ln_series_terms`][Self::ln_series_terms].
             pub const fn atanh_series(self, terms: $ue) -> Float<$f> {