+++ docs/changelog.md
@@ -151,6 +151,7 @@ Many feature gates are removed in order to make most features make always availa
   - types: `Cast`, `Compare`, `Cycle`, `CycleCount`, `False`, `Int`, `Interval`,  `NonExtreme*`, `NonValue*`, `Sign`, `True`.
   - traits: `ConstBool`.
 - new types: `IntAlloc`, `IntError`, `IntResult`, `NicheValueError`.
+- move float shared docs to `devela_base_num` prefixed with `_FLOAT_`.
 - move int shared docs to `devela_base_num` prefixed with `_INT_`.
 - update `Interval` to use individual `IncompatibleBounds` error.
 - un-gate `Compare` impls and many dependent const methods.